testsuite: Add TEST_OUTPUT to most verbose runnable tests

+++ test/runnable/template9.d
@@ -1040,7 +1040,7 @@ void test3467()
     a1 ~ a2; // line 7, Error
 }
 
-struct TS6806(size_t n) { pragma(msg, typeof(n)); }
+struct TS6806(uint n) { pragma(msg, typeof(n)); }
 static assert(is(TS6806!(1u) == TS6806!(1)));
 
 /**********************************/
